Making requested changes
Reducing the size of some columns in the template

Height is now a floatfield

Adding an indicator for age (months or years)

Data points are now highlighted

+++ drdown/medicalrecords/forms/curves_form.py
@@ -8,7 +8,7 @@ class CurvesForm(forms.ModelForm):
 
     weight = forms.FloatField(min_value=0.0)
 
-    height = forms.IntegerField(min_value=0)
+    height = forms.FloatField(min_value=0.0)
 
     cephalic_perimeter = forms.FloatField(min_value=0.0)
